Improved arbitrary generation for PPMs
Now header fields are given a range of valid values, and pixel values are generated within the proper range too. Validity-based generation was tried, but given how small the set of bytestrings that can be successfully parsed is, this method was extremely inefficient, despite it being nicer to define what it means to have a valid PPM bytestring. As well, Hedgehog was given a go, but I preferred the typeclass-based generation in this case, separating generation from the actual properties themselves.